*WIP* - Repository Deltas implementation start
git-svn-id: http://aptoide.org/repo/aptoide-client/unstable/v3@524 8d55eb69-6847-4b75-98a7-935b3913b0ec

+++ src/cm/aptoide/pt/data/xml/RepoBareParser.java
@@ -100,7 +100,7 @@ public class RepoBareParser extends DefaultHandler{
 //				Log.d("Aptoide-RepoBareParser", "app: "+application.getApplicationName()+", appHashid (Not full): "+application.getHashid()+", category: "+tagContentBuilder.toString().trim()+", categoryHashid: "+application.getCategoryHashid());
 				break;
 			case timestamp:
-				application.setTimestamp(Integer.parseInt(tagContentBuilder.toString()));
+				application.setTimestamp(Long.parseLong(tagContentBuilder.toString()));
 				break;
 //			case minSdk:	//TODO filters
 //				application.setVersionName(new String(chars).substring(start, start + length));
